+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - fixed
   - Fix throttling for fetched bookmarks
   - Fix `Cache::exists`
+  - Fix clean up for `Ctrl+C` shortcut
 - changed
   - Rename subcommand from `update` to `sync`
   - Rename flag from `--all` to `--replace` for `bogrep fetch`